Chapter 13 Fixes2
