+++ 13-managing-state/projects/final/lib/network/spoonacular_service.dart
@@ -11,11 +11,15 @@ part 'spoonacular_service.chopper.dart';
 const String apiKey = '<Add Your Key Here>';
 const String apiUrl = 'https://api.spoonacular.com/';
 
-typedef RecipeResponse = Response<Result<QueryResult>>;
-
 @ChopperApi()
 abstract class SpoonacularService extends ChopperService
     implements ServiceInterface {
+  /// Get the details of a specific recipe
+  @override
+  @Get(path: 'recipes/{id}/information?includeNutrition=false')
+  Future<RecipeDetailsResponse> queryRecipe(
+    @Path('id') String id,
+  );
 
   /// Get a list of recipes that match the query string
   @override
@@ -26,14 +30,6 @@ abstract class SpoonacularService extends ChopperService
     @Query('number') int number,
   );
 
-  /// Get the details of a specific recipe
-  @override
-  @Get(path: 'recipes/{id}/information?includeNutrition=false')
-  Future<Response<Result<Recipe>>> queryRecipe(
-    @Path('id') String id,
-  );
-
-  /// Create a new Chopper Client for our recipe service
   static SpoonacularService create() {
     final client = ChopperClient(
       baseUrl: Uri.parse(apiUrl),
