+++ 13-managing-state/projects/final/lib/ui/recipes/recipe_details.dart
@@ -1,11 +1,11 @@
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:cached_network_image/cached_network_image.dart';
-import 'package:chopper/chopper.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_html/flutter_html.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:lumberdash/lumberdash.dart';
+import '../../network/service_interface.dart';
 import '../../providers.dart';
 
 import '../../data/models/recipe.dart';
@@ -80,7 +80,7 @@ class _RecipeDetailsState extends ConsumerState<RecipeDetails> {
     );
   }
 
-  void readRecipe(AsyncSnapshot<Response<Result<Recipe>>> snapshot) {
+  void readRecipe(AsyncSnapshot<RecipeDetailsResponse> snapshot) {
     final result = snapshot.data?.body;
     if (result is Success<Recipe>) {
       final body = result.value;
