+++ 13-managing-state/projects/starter/lib/network/spoonacular_converter.dart
@@ -2,7 +2,6 @@ import 'dart:convert';
 import 'package:chopper/chopper.dart';
 import 'query_result.dart';
 
-import '../data/models/models.dart';
 import 'model_response.dart';
 import 'spoonacular_model.dart';
 
@@ -70,42 +69,4 @@ class SpoonacularConverter implements Converter {
     return decodeJson<BodyType, InnerType>(response);
   }
 
-}
-
-/// Methods to convert network recipes into local recipes
-List<Recipe> spoonacularResultsToRecipe(SpoonacularResults result) {
-  final recipes = <Recipe>[];
-  for (final result in result.results) {
-    recipes.add(spoonacularToRecipe(result));
-  }
-  return recipes;
-}
-
-Recipe spoonacularToRecipe(SpoonacularResult result) {
-  return Recipe(
-      id: result.id,
-      image: result.image,
-      label: result.title,
-      bookmarked: false,
-      ingredients: const <Ingredient>[],
-      description: result.title);
-}
-
-Recipe spoonacularRecipeToRecipe(SpoonacularRecipe spoonacularRecipe) {
-  final ingredients = <Ingredient>[];
-  for (final ingredient in spoonacularRecipe.extendedIngredients) {
-    ingredients.add(Ingredient(
-        id: ingredient.id,
-        name: ingredient.name,
-        amount: ingredient.amount,
-        recipeId: spoonacularRecipe.id));
-  }
-  return Recipe(
-    id: spoonacularRecipe.id,
-    label: spoonacularRecipe.title,
-    image: spoonacularRecipe.image,
-    bookmarked: false,
-    description: spoonacularRecipe.summary,
-    ingredients: ingredients,
-  );
 }
